refactor: Cron 메서드 scheduler로 이동
