+++ package.json
@@ -64,9 +64,10 @@
     "@nestjs/cli": "^10.0.0",
     "@nestjs/schematics": "^10.0.0",
     "@nestjs/testing": "^10.0.0",
-    "@types/cron": "^2.4.0",
     "@types/bcrypt": "^5.0.1",
+    "@types/cache-manager-ioredis": "^2.0.5",
     "@types/cookie-parser": "^1.4.5",
+    "@types/cron": "^2.4.0",
     "@types/express": "^4.17.20",
     "@types/jest": "^29.5.2",
     "@types/node": "^20.3.1",
