+++ src/feature/city/city.module.ts
@@ -1,8 +1,9 @@
-import { Module } from '@nestjs/common';
-import { CityService } from './city.service';
-import { CityController } from './city.controller';
 import { CacheModule } from '@nestjs/cache-manager';
+import { Module } from '@nestjs/common';
 import * as redisStore from 'cache-manager-ioredis';
+import { SchedulerModule } from '../scheduler/scheduler.module';
+import { CityController } from './city.controller';
+import { CityService } from './city.service';
 
 @Module({
   imports: [
@@ -13,6 +14,7 @@ import * as redisStore from 'cache-manager-ioredis';
       port: 6379,
       ttl: 100000,
     }),
+    SchedulerModule,
   ],
   controllers: [CityController],
   providers: [CityService],
