+++ src/feature/scheduler/scheduler.module.ts
@@ -2,9 +2,22 @@ import { Module } from '@nestjs/common';
 import { RestaurantModule } from '../restaurant/restaurant.module';
 import { SchedulerService } from './scheduler.service';
 import { ExternalApiModule } from '../externalApi/externalApi.module';
+import { CacheModule } from '@nestjs/cache-manager';
+import * as redisStore from 'cache-manager-ioredis';
 
 @Module({
-  imports: [RestaurantModule, ExternalApiModule],
+  imports: [
+    CacheModule.register({
+      isGlobal: true,
+      store: redisStore,
+      host: 'localhost',
+      port: 6379,
+      ttl: 100000,
+    }),
+    RestaurantModule,
+    ExternalApiModule,
+  ],
   providers: [SchedulerService],
+  exports: [SchedulerService],
 })
 export class SchedulerModule {}
